requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
agno>=2.4.0
openai>=1.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Desktop app dependencies
pystray>=0.19.0
Pillow>=10.0.0
